Extract github-related operations into their own file
